requirements.txt
numpy>=1.20.0
opencv-python>=4.5.0
open3d>=0.15.0
pyyaml>=6.0
torch>=1.13.0
torchvision>=0.14.0
pillow>=9.0.0
tqdm>=4.62.0